seperated navbar into 3 files in its own folder
