+++ client/style/navbar.scss
@@ -7,6 +7,7 @@
   background-color: rgb(121,0,0);
   color: white;
   text-align: center;
+  overflow: auto;
   &-logo {
     @include background-pic("https://vectr.com/mcabrex/gc2T4Hdhl.svg?width=640&height=640&select=gc2T4Hdhlpage0",contain);
     height: 15vh;
@@ -23,18 +24,19 @@
     font-size: 2vw;
   }
   &-items {
-    font-size: 2vw;
+    font-size: 1vw;
     display: grid;
     grid-template-rows: repeat(3, 5vh);
     margin: 5%;
     &-div {
       @extend %center-child;
+      &:hover {
+        background-color: rgb(29, 29, 29);
+        box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 1), 0 6px 20px 0 rgba(0, 0, 0, 1);
+      }
     }
     &-link {
-      @include button(white,black);
-      &:active {
-        background-color: black;
-      }
+      @include button(white,rgba(15, 15, 15, 0));
     }
   }
 }
